refactor: Split hash_join.rs into focused modules
Refactors the 888-line hash_join.rs into a modular structure for better
separation of concerns and easier optimization:

- build.rs: Hash table building (sequential + parallel strategies)
- inner.rs: INNER JOIN implementation
- outer.rs: LEFT OUTER JOIN implementation
- tests.rs: Comprehensive test suite
- mod.rs: Public API and utilities

Benefits:
- Clearer separation of sequential vs parallel strategies
- Easier to optimize each join type independently
- Better testing with focused unit tests per join type
- Simpler performance tuning and profiling

All existing tests pass. No performance regression.

Closes #2386
